Added Controller with Review Service and It's Implementation

+++ src/main/java/org/example/uberreviewservice/Services/ReviewServiceImpl.java
@@ -1,7 +1,9 @@
 package org.example.uberreviewservice.Services;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.example.uberreviewservice.Repositories.ReviewRepository;
 import org.example.uberreviewservice.models.Review;
+import org.hibernate.FetchNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,22 +19,60 @@ public class ReviewServiceImpl implements ReviewService{
     }
 
     @Override
-    public Optional<Review> findReviewById(Long id) {
-        return reviewRepository.findById(id);
-    }
-
-    @Override
-    public List<Review> findAllReviews() {
-        return reviewRepository.findAll();
-    }
-
-    @Override
-    public Boolean deleteReviewById(Long id) {
+    public Optional<Review> getReviewById(Long id) throws EntityNotFoundException {
+        Optional<Review> review ;
         try{
+            review = reviewRepository.findById(id);
+            if(review.isEmpty()){
+                throw new EntityNotFoundException("Review with id " + id + " not found");
+            }
+        }
+        catch(Exception e){
+            e.printStackTrace();
+            if(e.getClass() == EntityNotFoundException.class){
+                throw new FetchNotFoundException("Review with id " + id + " not found", id);
+            }
+            throw new FetchNotFoundException("Unable to fetch, PLease try again later!", id);
+        }
+        return review;
+    }
+
+    @Override
+    public List<Review> getAllReviews() {
+        return this.reviewRepository.findAll();
+    }
+
+    @Override
+    public boolean deleteReviewById(Long id) {
+        try {
+            // Check if the entity exists before deleting
+            if (!reviewRepository.existsById(id)) {
+                throw new EntityNotFoundException("Review with ID " + id + " not found.");
+            }
+
+            // Delete the entity
             reviewRepository.deleteById(id);
             return true;
-        }catch (Exception e){
+        } catch (Exception e) {
+            // Log the error if needed
             return false;
         }
+    }
+
+    @Override
+    public Review publishReview(Review request) {
+        return this.reviewRepository.save(request);
+    }
+
+    @Override
+    public Review updateReview(Long reviewId, Review request) {
+        Review review = this.reviewRepository.findById(reviewId).orElseThrow(EntityNotFoundException::new);
+        if(request.getRating()!=null){
+            review.setRating(request.getRating());
+        }
+        if(request.getContent()!=null){
+            review.setContent(request.getContent());
+        }
+        return this.reviewRepository.save(review);
     }
 }
